+++ CHANGELOG.md
@@ -5,11 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.0.3] - 2023-02-07
+## [0.0.3] - 2023-03-06
+
+### Fixed
+- Missing docstrings.
+- Removed docstrings from help messages.
 
 ### Change
-- Database directorys is now `~/.todi/`. 
+- Database directory is now `~/.todi/`. 
 
+### Added
+- New feature: Import and Export todi databases.
+    - Import from anywhere, by indicating the path of the backup file.
+    - Backup files are exported to `~/todi_backups/` directory. 
 
 
 ## [0.0.2] - 2022-09-20